% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_add_new_summoner_by_name}
\alias{db_add_new_summoner_by_name}
\title{Title}
\usage{
db_add_new_summoner_by_name(summoner = "Kewizard FR")
}
\arguments{
\item{summoner}{Summoner name.}
}
\value{
A logical value indicating whether the operation succeeded.
}
\description{
Title
}
\examples{
if (interactive()) {
  db_add_new_summoner_by_name()
}
}
