% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_add_or_update_summoner}
\alias{db_add_or_update_summoner}
\title{Summoner Data}
\usage{
db_add_or_update_summoner(summoner = "Kewizard FR")
}
\arguments{
\item{summoner}{Summoner name.}
}
\value{
A logical value indicating whether the operation succeeded.
}
\description{
Summoner Data
}
\examples{
if (interactive()) {
  db_add_or_update_summoner()
}
}
