% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{db_dir_create}
\alias{db_dir_create}
\title{Database Global Management}
\usage{
db_dir_create()
}
\value{
A logical scalar indicating whether the directory was created.
}
\description{
Database Global Management
}
\examples{
if (interactive()) {
  db_dir_create()
}
}
