% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{db_exists}
\alias{db_exists}
\alias{db_table_exists}
\title{Database Checks}
\usage{
db_exists()

db_table_exists(table)
}
\arguments{
\item{table}{Name of a database table.}
}
\value{
A logical scalar indicating whether the database exists.
}
\description{
Database Checks
}
\examples{
db_exists()
db_table_exists("summoners")
}
