% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{db_path}
\alias{db_path}
\alias{db_dir}
\title{Database Configuration}
\usage{
db_path()

db_dir()
}
\value{
Path to the SQLite database.
}
\description{
Database Configuration
}
\examples{
db_dir()
db_path()
}
