% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.db_summoner_name_exists}
\alias{.db_summoner_name_exists}
\title{Check Local Database}
\usage{
.db_summoner_name_exists(summoner = "Kewizard FR")
}
\arguments{
\item{summoner}{Summoner name.}
}
\value{
A logical scalar indicating whether the summoner name is present in the local database.
}
\description{
Check Local Database
}
