% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{.format_date_from_numeric}
\alias{.format_date_from_numeric}
\title{Format Dates}
\usage{
.format_date_from_numeric(date)
}
\arguments{
\item{date}{Date.}
}
\value{
Formatted date as string value.
}
\description{
Format Dates
}
