% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_summoner_by_name}
\alias{get_summoner_by_name}
\title{Get summoner data}
\usage{
get_summoner_by_name(
  summoner = "Kewizard FR",
  host = "euw1.api.riotgames.com",
  token = Sys.getenv("RIOT_API_KEY")
)
}
\arguments{
\item{summoner}{Summoner name.}

\item{host}{API host.}

\item{token}{API token.}
}
\value{
Content of the reply.
}
\description{
Get summoner data
}
\examples{
get_summoner_by_name()
}
